auto tagging by system

+++ app/commands/create_tag.rb
@@ -4,12 +4,13 @@ class CreateTag
   prepend SimpleCommand
   include ActiveModel::Validations
 
-  attr_reader :bookmark, :tag_names, :user
+  attr_reader :bookmark, :tag_names, :user, :with_remove
 
-  def initialize(bookmark, tag_names, user)
+  def initialize(bookmark, tag_names, user, with_remove = true)
     @bookmark = bookmark
     @tag_names = tag_names
     @user = user
+    @with_remove = with_remove
   end
 
   def call
@@ -18,7 +19,7 @@ class CreateTag
     add_tags = (new_tags - old_tags)
     remove_tags = (old_tags - new_tags)
     add_tags.each { |x| bookmark.taggings.create(tag: x, user: user) }
-    bookmark.taggings.where(tag: remove_tags).where(user: user).destroy_all
+    bookmark.taggings.where(tag: remove_tags).where(user: user).destroy_all if with_remove
     bookmark.tags.where(taggings: { user: user }).reload
   end
 end
